# docs/api-spec.yml
openapi: 3.0.0
info:
  title: PHYLOViZ Web Platform
  description: A Modular and Web-Based Tool for Phylogenetic Analysis.
  version: 1.0.0
servers:
  - url: https://web.phyloviz.net
    description: Production server
  - url: http://localhost:8083
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: [ ]
...
paths:
...
  /projects/{projectId}/files/isolate-data:
    post:
      tags:
        - FileTransfer > Projects > Files > Isolate Data
      summary: Upload Isolate Data File
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      parameters:
        - name: projectId
          in: path
          schema:
            type: string
          required: true
      responses:
        '201':
          description: Created
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                projectId: ''
                isolateDataId: ''
  /projects/{projectId}/files/isolate-data/{isolateDataId}:
    delete:
      tags:
        - Administration > Projects > Files > Isolate Data
      summary: Delete Isolate Data File
      parameters:
        - name: projectId
          in: path
          schema:
            type: string
          required: true
        - name: isolateDataId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                projectId: ''
                isolateDataId: ''
    patch:
      tags:
        - Administration > Projects > Files > Isolate Data
      summary: Update Isolate Data File
      parameters:
        - name: projectId
          in: path
          schema:
            type: string
          required: true
        - name: isolateDataId
          in: path
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: ''
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
                example:
                  name: ''
    get:
      tags:
        - Visualization > Files > Isolate Data
      summary: Get Isolate Data Keys
      parameters:
        - name: projectId
          in: path
          schema:
            type: string
          required: true
        - name: isolateDataId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                keys:
                  - '1'
                  - '2'
                  - '3'
  /projects/{projectId}/files/isolate-data/{isolateDataId}/file:
    get:
      tags:
        - FileTransfer > Projects > Files > Isolate Data
      summary: Download Isolate Data File
      parameters:
        - name: projectId
          in: path
          schema:
            type: string
          required: true
        - name: isolateDataId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/octet-stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...